// src/generated/local-layout.ts
export const localLayout = {
  "local": {
    "block": "local",
    "file": "LOW.FOR",
    "routine": "LOW",
    "address": 224,
    "words": 200,
    "declaredWords": 200,
    "mapLine": 22,
    "fields": {
      "locdum": {
        "offset": 0,
        "words": 200,
        "dimensions": [
          {
            "lower": 1,
            "length": 200
          }
        ],
        "type": "implicit-integer",
        "typeSource": {
          "file": "PARAM.FOR",
          "line": 21
        },
        "file": "LOW.FOR",
        "line": 24
      }
    }
  },
// ...
} as const;